Rename helper module to fixture to avoid ambiguous module names.
